// src/pages/mbb-documents/internal-outbound/interfaces/internal-outbound-mbb.interface.ts
export interface InternalOutboundMbbDocument {
  _id: string;
  reg_num: string;
  reg_date: string;
  journal_index: string;
  doc_num: string;
  doc_date: string;
  recipient: string;
  summary: string;
  response_reference_number: string;
  internal_inbound_id: string;
  assignee: string;
  deadline: string;
  status: string;
}